node to page: Final TODO-fixes
Updates #2297
